// TankMateGasApp/src/app/interface/estacionInterna.interface.ts
export interface EstacionInterna {
  cp?: string;
  direccion?: string;
  horario?: string;
  latitud?: string;
  localidad?: string;
  longitud?: string;
  margen?: string;
  municipio?: string;
  provincia?: string;
  rotulo?: string;
  tipoVenta?: string;
  precioAdblue?: string;
  precioAmoniaco?: string;
  precioBiodiesel?: string;
  precioBioetanol?: string;
  precioGasoleoA?: string;
  precioGasoleoB?: string;
  precioGasolina95E5?: string;
  precioGasolina98E5?: string;
  porcentajeBioEtanol?: string;
  porcentajeEster?: string;
  ideess?: string;
  idMunicipio?: string;
  idProvincia?: string;
  idCCAA?: string;
}